vendor: create initial infrastrure for linux vendoring
Since our node library only uses a fraction of the kernel there is no
point including all the files and all its history. Therefore, we clone
it and build it while marking the files that were used during
compilation. We delete anything that wasn't used and then we vendor the
code.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>