alternative syscall implementation
This commit removes the previous implementation that autogenerated C++
bindings for every syscall and instead uses a single, generic binding
(NAN_METHOD(syscall) and SyscallWorker) that accepts a syscall number.

On the JS side, syscalls are defined in lib/syscall.js and are sync or
async depending on whether or not the callback parameter is null.

This commit also includes a copy of the `fs` module implementation which
has its process.binding('fs') replaced with lib/binding.js.

By implementing the few methods required in lib/binding.js we
automatically get all the standard filesystem API like stream support.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>