expose more of LKL in nodejs world
With this commit a few things that were happening automatically now need
to be explicitly done from nodeJS. Namely, startKernel and haltKernel
are no longer called implicitly by mount.

Also, the kernel will no longer chroot to the mountpoint of the first
mount. Instead, the mountpoint is returned to the user of the libary and
needs to be prepended to any path before passing it to lkl.

The two things above allow multiple images to be mounted simultaneously
with the same kernel instance.

The filesystem type option was also added to the lkl.mount() call to
enable mounting of filesystems other than ext4.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>