native: implement async js disk operations
To support js disk operations the code flow needs to go through a few
hoops that I'll try to explain here.

When a disk is added to the lkl kernel, an `lkl_dev_blk_ops` structure
is associated with it that defines the functions that will serve I/O
requests when the filesystem needs them. The problem is that those
functions will be invoked in a lkl thread and therefore cannot call any
JS functions.

Let's follow a full cycle of a call to ops.request().

First, `js_request_entry()` is called in a lkl thread by the kernel. This
function has to fulfill the request *synchronously*. To do that, it
creates a semaphore(`js_sem`) and runs `request()` in the main loop using
`run_in_default_loop()`. After that call it waits on the semaphore
indefinitely until the some other thread posts to it.

Now `request()`, which is running on the main loop, can call the JS
function (`js_ops->request()`) that was supplied by the user. This is
non-blocking and `request()` will return as soon as the user's function
(which is also async) returns. In order for the native side to be
notified when the user has finished processing the request, `request()`
passes a callback that is associated with a C++ function
(`request_done`) and a state. The state includes the original `js_sem`.

At some point, the user is ready to return control to lkl and calls the
supplied callback. This causes `request_done()` to run. This function
will process the user's data passed in the callback and finally do a
`uv_sem_post(&js_sem)`. This unblocks the `js_request_entry()` call and
control is returned to lkl.

There are two noteworthy things to mention about this process. Since the
original `js_request_entry()` call is waiting for all the others to
finish before returning, we can store all the state in the stack and
pass pointers pointing to its stack around. This avoids heap allocations
and having to manually free memory. A side effect of this is that we
have to ensure that the callback is not called twice, since the second
time the state will be invalid and will be pointing to invalid memory.
To accomplish that the callback self-destructs its state the first time
it runs.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>